+++ idlang/src/main/java/mimuw/idlearn/idlang/logic/base/Type.java
@@ -1,7 +1,7 @@
 package mimuw.idlearn.idlang.logic.base;
 
 public enum Type {
-	Integer,
+	Long,
 	Table,
 	Null
 }
