+++ idlang/src/main/java/mimuw/idlearn/idlang/logic/exceptions/OutOfBoundsException.java
@@ -1,7 +1,7 @@
 package mimuw.idlearn.idlang.logic.exceptions;
 
 public class OutOfBoundsException extends SimulationException {
-	public OutOfBoundsException(int index, int size) {
+	public OutOfBoundsException(long index, long size) {
 		super("Index " + index + " out of bounds for array of size " + size + ".");
 	}
 }
