+++ scoring/src/test/java/mimuw/idlearn/scoring/TestRunnerTest.java
@@ -9,7 +9,6 @@ import mimuw.idlearn.idlang.logic.operators.TwoArgOperator;
 import mimuw.idlearn.packages.PackageManager;
 import mimuw.idlearn.packages.ProblemPackage;
 import java.io.IOException;
-import java.util.concurrent.Semaphore;
 
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.*;
@@ -22,9 +21,9 @@ public class TestRunnerTest {
 		assertEquals(pack.getTitle(), "Addition");
 
 		Expression solution = new Block(
-				new InputHandler(new Variable(Type.Integer, "x"), new Variable(Type.Integer, "y")),
-				new Assignment("x", TwoArgOperator.newAdd(new Variable(Type.Integer, "x"), new Variable(Type.Integer, "y")), false),
-				new OutputHandler(new Variable(Type.Integer, "x"))
+				new InputHandler(new Variable(Type.Long, "x"), new Variable(Type.Long, "y")),
+				new Assignment("x", TwoArgOperator.newAdd(new Variable(Type.Long, "x"), new Variable(Type.Long, "y")), false),
+				new OutputHandler(new Variable(Type.Long, "x"))
 		);
 
 		TestRunner testRunner = new TestRunner(pack, solution);
